Refactor the peak limiter (output is bit-exact)
